rublk: abstract generic offload handler
This commit introduces a generic offload handling mechanism to simplify
the implementation of ublk targets that need to offload I/O operations
to worker threads.

The key changes are:

- **`src/offload/handler.rs`**: A new `OffloadHandler` struct is
  introduced. This struct encapsulates the logic for managing a worker
  thread, an eventfd for signaling, and the communication channels
  (Sender/Receiver) for jobs and completions. It provides a generic
  `setup_worker_thread` function that can be used by any target.

- **`src/offload/mod.rs`**: The `OffloadTargetLogic` trait is defined to
  provide a common interface for targets that use the offload
  mechanism. It requires the implementation of `setup_read_worker`,
  `setup_flush_worker`, and `handle_io`.

- **`src/compress.rs`**: The `compress` target is refactored to use the
  new offload handler. The `QueueHandler` is simplified, and the logic
  for setting up worker threads and handling offloaded I/O is now
  delegated to the `OffloadHandler`. This makes the `compress` target's
  code cleaner and more focused on its specific logic (i.e., RocksDB
  operations).

In Rust, this is implemented by:

- Using generic types (`<Job, C, F>`) in `setup_worker_thread` to make
  it adaptable to different types of jobs and completion data.
- Leveraging traits (`OffloadTargetLogic`) to define a contract for
  offloadable targets, promoting polymorphism and code reuse.
- Using `mpsc::channel` for communication between the main thread and
  the worker threads, which is a standard and safe way to handle
  concurrent operations in Rust.
- The `nix` crate is used for creating and managing the `eventfd`, which
  is a low-level Linux feature for event notification.

This refactoring improves the overall design by separating the generic
offload mechanism from the specific implementation details of each
target, leading to more maintainable and reusable code.
