+++ src/compress.rs
@@ -1,10 +1,11 @@
+use crate::offload::{
+    handler::{setup_worker_thread, Completion, FlushJob, ReadJob},
+    OffloadTargetLogic,
+};
 use libublk::{
     ctrl::UblkCtrl,
-    helpers::IoBuf,
     io::{UblkDev, UblkIOCtx, UblkQueue},
-    UblkIORes,
 };
-use nix::sys::eventfd::{eventfd, EfdFlags};
 use rocksdb::{
     ColumnFamily, ColumnFamilyDescriptor, DBCompressionType, Options, SliceTransform, WriteBatch,
     WriteOptions, DB,
@@ -14,7 +15,7 @@ use serde_json;
 use std::fs::{File, OpenOptions};
 use std::io::Write;
 use std::path::PathBuf;
-use std::sync::mpsc::{channel, Receiver, Sender};
+use std::sync::mpsc::{Receiver, Sender};
 use std::sync::Arc;
 
 #[derive(clap::Args, Debug, Clone)]
@@ -48,24 +49,6 @@ struct CompressJson {
     logical_block_size: u32,
     physical_block_size: u32,
 }
-
-struct ReadJob {
-    tag: u16,
-    start_sector: u64,
-    nr_sectors: u32,
-    buf_addr: u64,
-}
-
-struct FlushJob {
-    tag: u16,
-}
-
-struct Completion {
-    tag: u16,
-    result: Result<i32, i32>,
-}
-
-const POLL_TAG: u16 = u16::MAX;
 
 fn handle_read(
     db: &DB,
@@ -144,223 +127,67 @@ fn handle_discard(
     Ok(0)
 }
 
-struct OffloadHandler<'a, Job> {
-    q: &'a UblkQueue<'a>,
-    bufs_ptr: *mut IoBuf<u8>,
-    efd: i32,
-    job_tx: Sender<Job>,
-    completion_rx: Receiver<Completion>,
-    poll_op: u32,
-}
-
-impl<'a, Job> OffloadHandler<'a, Job> {
-    fn new(
-        q: &'a UblkQueue<'a>,
-        bufs_ptr: *mut IoBuf<u8>,
-        efd: i32,
-        job_tx: Sender<Job>,
-        completion_rx: Receiver<Completion>,
-        poll_op: u32,
-    ) -> Self {
-        let handler = Self {
-            q,
-            bufs_ptr,
-            efd,
-            job_tx,
-            completion_rx,
-            poll_op,
-        };
-        handler.submit_poll_sqe();
-        handler
-    }
-
-    fn submit_poll_sqe(&self) {
-        let user_data = UblkIOCtx::build_user_data(POLL_TAG, self.poll_op, 0, true);
-        let sqe = io_uring::opcode::PollAdd::new(io_uring::types::Fd(self.efd), libc::POLLIN as _)
-            .build()
-            .user_data(user_data);
-        self.q.ublk_submit_sqe_sync(sqe).unwrap();
-    }
-
-    fn handle_completion(&self) {
-        let mut buf = [0u8; 8];
-        nix::unistd::read(self.efd, &mut buf).unwrap();
-
-        while let Ok(completion) = self.completion_rx.try_recv() {
-            let result = match completion.result {
-                Ok(r) => UblkIORes::Result(r),
-                Err(e) => UblkIORes::Result(e),
-            };
-            let tag = completion.tag;
-            let buf_addr = unsafe {
-                let io_buf: &mut IoBuf<u8> = &mut *self.bufs_ptr.add(tag as usize);
-                io_buf.as_mut().as_mut_ptr()
-            };
-            self.q.complete_io_cmd(tag, buf_addr, Ok(result));
-        }
-        self.submit_poll_sqe();
-    }
-}
-
-impl<'a> OffloadHandler<'a, ReadJob> {
-    fn send_job(&self, tag: u16, iod: &libublk::sys::ublksrv_io_desc, buf: &mut [u8]) {
-        self.job_tx
-            .send(ReadJob {
-                tag,
-                start_sector: iod.start_sector,
-                nr_sectors: iod.nr_sectors,
-                buf_addr: buf.as_mut_ptr() as u64,
-            })
-            .unwrap();
-    }
-}
-
-impl<'a> OffloadHandler<'a, FlushJob> {
-    fn send_job(&self, tag: u16) {
-        self.job_tx.send(FlushJob { tag }).unwrap();
-    }
-}
-
-struct QueueHandler<'a> {
-    q: &'a UblkQueue<'a>,
-    db: &'a Arc<DB>,
-    bufs_ptr: *mut IoBuf<u8>,
-    max_buf_len: usize,
+struct CompressTarget {
+    db: Arc<DB>,
     lbs: u32,
     read_only: bool,
-    read_handler: OffloadHandler<'a, ReadJob>,
-    flush_handler: OffloadHandler<'a, FlushJob>,
 }
 
-impl<'a> QueueHandler<'a> {
-    fn setup_worker_thread<Job, C, F>(efd: i32, handler: F) -> (Sender<Job>, Receiver<C>)
-    where
-        Job: Send + 'static,
-        C: Send + 'static,
-        F: Fn(Job) -> C + Send + 'static,
-    {
-        let (job_tx, job_rx) = channel();
-        let (completion_tx, completion_rx) = channel();
-
-        std::thread::spawn(move || {
-            for job in job_rx {
-                let completion = handler(job);
-                if completion_tx.send(completion).is_err() {
-                    break;
-                }
-                nix::unistd::write(efd, &1u64.to_le_bytes()).unwrap();
+impl OffloadTargetLogic for CompressTarget {
+    fn setup_read_worker(
+        &self,
+        efd: i32,
+    ) -> (Sender<ReadJob>, Receiver<Completion>) {
+        let db = self.db.clone();
+        let lbs = self.lbs;
+        setup_worker_thread(efd, move |job: ReadJob| {
+            let buf_slice = unsafe {
+                std::slice::from_raw_parts_mut(
+                    job.buf_addr as *mut u8,
+                    (job.nr_sectors << 9) as usize,
+                )
+            };
+            let res = handle_read(&db, job.start_sector, job.nr_sectors, buf_slice, lbs);
+            Completion {
+                tag: job.tag,
+                result: res,
             }
-        });
-
-        (job_tx, completion_rx)
+        })
     }
 
-    fn new(
-        q: &'a UblkQueue<'a>,
-        db: &'a Arc<DB>,
-        bufs: &'a mut [IoBuf<u8>],
-        dev: &'a UblkDev,
-    ) -> Self {
-        let flush_efd = eventfd(0, EfdFlags::EFD_CLOEXEC).unwrap();
-        let read_efd = eventfd(0, EfdFlags::EFD_CLOEXEC).unwrap();
-        let lbs = 1u32 << dev.tgt.params.basic.logical_bs_shift;
-        let bufs_ptr = bufs.as_mut_ptr();
-
-        let db_read = db.clone();
-        let (read_job_tx, read_completion_rx) =
-            Self::setup_worker_thread(read_efd, move |job: ReadJob| {
-                let buf_slice = unsafe {
-                    std::slice::from_raw_parts_mut(
-                        job.buf_addr as *mut u8,
-                        (job.nr_sectors << 9) as usize,
-                    )
-                };
-                let res = handle_read(&db_read, job.start_sector, job.nr_sectors, buf_slice, lbs);
-                Completion {
-                    tag: job.tag,
-                    result: res,
-                }
-            });
-
-        let db_flush = db.clone();
-        let (flush_job_tx, flush_completion_rx) =
-            Self::setup_worker_thread(flush_efd, move |job: FlushJob| {
-                let res = if let Err(e) = db_flush.flush() {
-                    log::error!("Background rocksdb flush error: {}", e);
-                    Err(-libc::EIO)
-                } else {
-                    Ok(0)
-                };
-                Completion {
-                    tag: job.tag,
-                    result: res,
-                }
-            });
-
-        let read_handler = OffloadHandler::new(
-            q,
-            bufs_ptr,
-            read_efd,
-            read_job_tx,
-            read_completion_rx,
-            libublk::sys::UBLK_IO_OP_READ,
-        );
-
-        let flush_handler = OffloadHandler::new(
-            q,
-            bufs_ptr,
-            flush_efd,
-            flush_job_tx,
-            flush_completion_rx,
-            libublk::sys::UBLK_IO_OP_FLUSH,
-        );
-
-        Self {
-            q,
-            db,
-            bufs_ptr,
-            max_buf_len: dev.dev_info.max_io_buf_bytes as usize,
-            lbs,
-            read_only: (dev.tgt.params.basic.attrs & libublk::sys::UBLK_ATTR_READ_ONLY) != 0,
-            read_handler,
-            flush_handler,
-        }
+    fn setup_flush_worker(
+        &self,
+        efd: i32,
+    ) -> (Sender<FlushJob>, Receiver<Completion>) {
+        let db = self.db.clone();
+        setup_worker_thread(efd, move |job: FlushJob| {
+            let res = if let Err(e) = db.flush() {
+                log::error!("Background rocksdb flush error: {}", e);
+                Err(-libc::EIO)
+            } else {
+                Ok(0)
+            };
+            Completion {
+                tag: job.tag,
+                result: res,
+            }
+        })
     }
 
-    fn handle_io(&self, tag: u16, io_ctx: &UblkIOCtx) {
-        if io_ctx.is_tgt_io() {
-            match UblkIOCtx::user_data_to_op(io_ctx.user_data()) {
-                libublk::sys::UBLK_IO_OP_FLUSH => {
-                    self.flush_handler.handle_completion();
-                    return;
-                }
-                libublk::sys::UBLK_IO_OP_READ => {
-                    self.read_handler.handle_completion();
-                    return;
-                }
-                _ => {}
-            }
-        }
-
-        let iod = self.q.get_iod(tag);
+    fn handle_io(
+        &self,
+        _q: &UblkQueue,
+        _tag: u16,
+        iod: &libublk::sys::ublksrv_io_desc,
+        buf: &mut [u8],
+    ) -> Result<i32, i32> {
         let op = iod.op_flags & 0xff;
-
-        let buf = unsafe {
-            let buf_len = std::cmp::min((iod.nr_sectors << 9) as usize, self.max_buf_len);
-            let io_buf: &mut IoBuf<u8> = &mut *self.bufs_ptr.add(tag as usize);
-            &mut io_buf.as_mut()[..buf_len]
-        };
-
-        let res = match op {
-            libublk::sys::UBLK_IO_OP_READ => {
-                self.read_handler.send_job(tag, iod, buf);
-                return;
-            }
+        match op {
             libublk::sys::UBLK_IO_OP_WRITE => {
                 if self.read_only {
                     Err(-libc::EACCES)
                 } else {
-                    handle_write(self.db, iod.start_sector, iod.nr_sectors, buf, self.lbs)
+                    handle_write(&self.db, iod.start_sector, iod.nr_sectors, buf, self.lbs)
                 }
             }
             libublk::sys::UBLK_IO_OP_DISCARD => {
@@ -368,21 +195,11 @@ impl<'a> QueueHandler<'a> {
                     Err(-libc::EACCES)
                 } else {
                     let cf = self.db.cf_handle("default").unwrap();
-                    handle_discard(self.db, cf, iod.start_sector, iod.nr_sectors, self.lbs)
+                    handle_discard(&self.db, cf, iod.start_sector, iod.nr_sectors, self.lbs)
                 }
             }
-            libublk::sys::UBLK_IO_OP_FLUSH => {
-                self.flush_handler.send_job(tag);
-                return;
-            }
             _ => Err(-libc::EINVAL),
-        };
-
-        let result = match res {
-            Ok(r) => UblkIORes::Result(r),
-            Err(e) => UblkIORes::Result(e),
-        };
-        self.q.complete_io_cmd(tag, buf.as_mut_ptr(), Ok(result));
+        }
     }
 }
 
@@ -393,7 +210,12 @@ fn q_sync_fn(qid: u16, dev: &UblkDev, db: &Arc<DB>) {
         .regiser_io_bufs(Some(&bufs))
         .submit_fetch_commands(Some(&bufs));
 
-    let handler = QueueHandler::new(&q, db, &mut bufs, dev);
+    let target = CompressTarget {
+        db: db.clone(),
+        lbs: 1u32 << dev.tgt.params.basic.logical_bs_shift,
+        read_only: (dev.tgt.params.basic.attrs & libublk::sys::UBLK_ATTR_READ_ONLY) != 0,
+    };
+    let handler = crate::offload::handler::QueueHandler::new(&q, &target, &mut bufs, dev);
 
     let io_handler = |_q: &UblkQueue, tag: u16, io_ctx: &UblkIOCtx| {
         handler.handle_io(tag, io_ctx);
